requirements.txt
tensorflow==2.14.0
scikit-learn
numpy
matplotlib
pillow
python-dotenv
streamlit
openai
dotenv